refactor dbms specific code
