+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/app/SqlClientBasicFactoryDefault.java
@@ -9,20 +9,22 @@ package de.ii.xtraplatform.features.sql.app;
 
 import de.ii.xtraplatform.features.domain.ConnectorFactory;
 import de.ii.xtraplatform.features.sql.domain.ConnectionInfoSql;
-import de.ii.xtraplatform.features.sql.domain.ConnectionInfoSql.Dialect;
 import de.ii.xtraplatform.features.sql.domain.SqlClientBasic;
 import de.ii.xtraplatform.features.sql.domain.SqlClientBasicFactory;
 import de.ii.xtraplatform.features.sql.domain.SqlConnector;
+import de.ii.xtraplatform.features.sql.domain.SqlDbmsAdapter;
+import de.ii.xtraplatform.features.sql.domain.SqlDbmsAdapters;
 import de.ii.xtraplatform.features.sql.domain.SqlDialect;
-import de.ii.xtraplatform.features.sql.domain.SqlDialectGpkg;
-import de.ii.xtraplatform.features.sql.domain.SqlDialectPostGis;
 import java.sql.Connection;
 
 public class SqlClientBasicFactoryDefault implements SqlClientBasicFactory {
   private final ConnectorFactory connectorFactory;
+  private final SqlDbmsAdapters dbmsAdapters;
 
-  public SqlClientBasicFactoryDefault(ConnectorFactory connectorFactory) {
+  public SqlClientBasicFactoryDefault(
+      ConnectorFactory connectorFactory, SqlDbmsAdapters dbmsAdapters) {
     this.connectorFactory = connectorFactory;
+    this.dbmsAdapters = dbmsAdapters;
   }
 
   @Override
@@ -47,7 +49,10 @@ public class SqlClientBasicFactoryDefault implements SqlClientBasicFactory {
       throw connectionError;
     }
 
-    return new SqlClientBasicDefault(connector, connectionInfo);
+    return new SqlClientBasicDefault(
+        connector,
+        dbmsAdapters.get(connectionInfo.getDialect()),
+        dbmsAdapters.getDialect(connectionInfo.getDialect()));
   }
 
   @Override
@@ -59,14 +64,14 @@ public class SqlClientBasicFactoryDefault implements SqlClientBasicFactory {
 
   private static class SqlClientBasicDefault implements SqlClientBasic {
     private final SqlConnector connector;
+    private final SqlDbmsAdapter dbmsAdapter;
     private final SqlDialect dialect;
 
-    private SqlClientBasicDefault(SqlConnector connector, ConnectionInfoSql connectionInfo) {
+    private SqlClientBasicDefault(
+        SqlConnector connector, SqlDbmsAdapter dbmsAdapter, SqlDialect dialect) {
       this.connector = connector;
-      this.dialect =
-          connectionInfo.getDialect() == Dialect.GPKG
-              ? new SqlDialectGpkg()
-              : new SqlDialectPostGis();
+      this.dbmsAdapter = dbmsAdapter;
+      this.dialect = dialect;
     }
 
     @Override
@@ -77,6 +82,11 @@ public class SqlClientBasicFactoryDefault implements SqlClientBasicFactory {
     @Override
     public SqlDialect getSqlDialect() {
       return dialect;
+    }
+
+    @Override
+    public SqlDbmsAdapter getDbmsAdapter() {
+      return dbmsAdapter;
     }
   }
 }
