+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/domain/ConnectionInfoSql.java
@@ -57,8 +57,8 @@ public interface ConnectionInfoSql extends ConnectionInfo {
    * @default PGIS
    */
   @Value.Default
-  default Dialect getDialect() {
-    return Dialect.PGIS;
+  default String getDialect() {
+    return SqlDbmsPgis.ID;
   }
 
   /**
