+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/domain/ConstantsResolver.java
@@ -5,7 +5,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package de.ii.xtraplatform.features.sql.app;
+package de.ii.xtraplatform.features.sql.domain;
 
 import de.ii.xtraplatform.features.domain.FeatureSchema;
 import de.ii.xtraplatform.features.domain.ImmutableFeatureSchema;
