+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/domain/SqlDbmsAdapters.java
@@ -1,0 +1,17 @@
+/*
+ * Copyright 2024 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.xtraplatform.features.sql.domain;
+
+public interface SqlDbmsAdapters {
+
+  boolean isSupported(String dbmsId);
+
+  SqlDbmsAdapter get(String dbmsId);
+
+  SqlDialect getDialect(String dbmsId);
+}
