+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/domain/SqlDbmsPgis.java
@@ -1,0 +1,13 @@
+/*
+ * Copyright 2024 interactive instruments GmbH
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+package de.ii.xtraplatform.features.sql.domain;
+
+public interface SqlDbmsPgis {
+
+  String ID = "PGIS";
+}
