+++ xtraplatform-features-sql/src/main/java/de/ii/xtraplatform/features/sql/domain/SqlDialectPgis.java
@@ -13,7 +13,6 @@ import static de.ii.xtraplatform.cql.domain.ArrayFunction.A_EQUALS;
 import static de.ii.xtraplatform.cql.domain.ArrayFunction.A_OVERLAPS;
 
 import com.google.common.base.Splitter;
-import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import de.ii.xtraplatform.cql.domain.ArrayFunction;
 import de.ii.xtraplatform.cql.domain.SpatialFunction;
@@ -22,23 +21,19 @@ import de.ii.xtraplatform.crs.domain.BoundingBox;
 import de.ii.xtraplatform.crs.domain.EpsgCrs;
 import de.ii.xtraplatform.features.domain.SchemaBase.Type;
 import de.ii.xtraplatform.features.sql.domain.SchemaSql.PropertyTypeInfo;
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
+import java.text.Collator;
 import java.time.Instant;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
-import java.util.LinkedHashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
 import java.util.Set;
-import org.immutables.value.Value;
 import org.threeten.extra.Interval;
 
-public class SqlDialectPostGis implements SqlDialect {
+public class SqlDialectPgis implements SqlDialect {
 
   private static final Splitter BBOX_SPLITTER =
       Splitter.onPattern("[(), ]").omitEmptyStrings().trimResults();
@@ -286,86 +281,12 @@ public class SqlDialectPostGis implements SqlDialect {
     return value.replaceAll("'", "''");
   }
 
+  /* NOTE: If the db uses e.g. the DE collation and some sort key actually contains e.g. umlauts
+           this might lead to wrong results.
+           To cover such cases, the locale would need to be configurable.
+  */
   @Override
-  public String geometryInfoQuery(Map<String, String> dbInfo) {
-    return String.format(
-        "SELECT f_table_schema AS \"%s\", f_table_name AS \"%s\", f_geometry_column AS \"%s\", coord_dimension AS \"%s\", srid AS \"%s\", type AS \"%s\" FROM geometry_columns;",
-        GeoInfo.SCHEMA,
-        GeoInfo.TABLE,
-        GeoInfo.COLUMN,
-        GeoInfo.DIMENSION,
-        GeoInfo.SRID,
-        GeoInfo.TYPE);
-  }
-
-  @Override
-  public Map<String, GeoInfo> getGeoInfo(Connection connection, DbInfo dbInfo) throws SQLException {
-    if (!(dbInfo instanceof DbInfoPgis)) {
-      throw new SQLException("Not a valid spatial PostgreSQL database.");
-    }
-    String query =
-        String.format(
-            "SELECT f_table_schema AS \"%s\", f_table_name AS \"%s\", f_geometry_column AS \"%s\", coord_dimension AS \"%s\", srid AS \"%s\", type AS \"%s\" FROM geometry_columns;",
-            GeoInfo.SCHEMA,
-            GeoInfo.TABLE,
-            GeoInfo.COLUMN,
-            GeoInfo.DIMENSION,
-            GeoInfo.SRID,
-            GeoInfo.TYPE);
-
-    Statement stmt = connection.createStatement();
-    ResultSet rs = stmt.executeQuery(query);
-    Map<String, GeoInfo> result = new LinkedHashMap<>();
-
-    while (rs.next()) {
-      result.put(
-          rs.getString(GeoInfo.TABLE),
-          ImmutableGeoInfo.of(
-              rs.getString(GeoInfo.SCHEMA),
-              rs.getString(GeoInfo.TABLE),
-              rs.getString(GeoInfo.COLUMN),
-              rs.getString(GeoInfo.DIMENSION),
-              rs.getString(GeoInfo.SRID),
-              forceAxisOrder(dbInfo).name(),
-              rs.getString(GeoInfo.TYPE)));
-    }
-
-    return result;
-  }
-
-  @Override
-  public DbInfo getDbInfo(Connection connection) throws SQLException {
-    String query = "SELECT version(), PostGIS_Lib_Version();";
-
-    Statement stmt = connection.createStatement();
-    ResultSet rs = stmt.executeQuery(query);
-    rs.next();
-
-    return ImmutableDbInfoPgis.of(rs.getString(1), rs.getString(2));
-  }
-
-  @Value.Immutable
-  interface DbInfoPgis extends DbInfo {
-
-    @Value.Parameter
-    String getPostgresVersion();
-
-    @Value.Parameter
-    String getPostGisVersion();
-  }
-
-  @Override
-  public List<String> getSystemSchemas() {
-    return ImmutableList.of("information_schema", "pg_catalog", "tiger", "tiger_data", "topology");
-  }
-
-  @Override
-  public List<String> getSystemTables() {
-    return ImmutableList.of(
-        "spatial_ref_sys",
-        "geography_columns",
-        "geometry_columns",
-        "raster_columns",
-        "raster_overviews");
+  public Collator getRowSortingCollator() {
+    return Collator.getInstance(Locale.US);
   }
 }
